database/connect_db.py
from dotenv import load_dotenv
import os
import psycopg2

load_dotenv()


def connect_database():
    connect = psycopg2.connect(
        database=os.getenv("DATABASE_NAME"),
        user=os.getenv("DATABASE_USER"),
        password=os.getenv("DATABASE_PASSWORD"),
        host=os.getenv("DATABASE_HOST"),
        port=os.getenv("DATABASE_PORT"),
    )

    return connect
